segmentar-sam-app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart
pillow
numpy
opencv-python
torch>=2.1.0
segment-anything
